Fixed the tests discovery issue due to filename.
